Attempt to fix the video not working on saucelabs, and some renaming of files
